+++ lib/routes/home/home_tab.dart
@@ -2,12 +2,15 @@ import 'package:flow/data/transactions_filter.dart';
 import 'package:flow/entity/transaction.dart';
 import 'package:flow/objectbox.dart';
 import 'package:flow/objectbox/actions.dart';
+import 'package:flow/prefs.dart';
+import 'package:flow/utils/optional.dart';
 import 'package:flow/widgets/default_transaction_filter_head.dart';
 import 'package:flow/widgets/general/wavy_divider.dart';
 import 'package:flow/widgets/grouped_transaction_list.dart';
 import 'package:flow/widgets/home/greetings_bar.dart';
 import 'package:flow/widgets/home/home/no_transactions.dart';
 import 'package:flow/widgets/home/transactions_date_header.dart';
+import 'package:flow/widgets/utils/time_and_range.dart';
 import 'package:flutter/material.dart';
 import 'package:moment_dart/moment_dart.dart';
 
@@ -21,17 +24,29 @@ class HomeTab extends StatefulWidget {
 }
 
 class _HomeTabState extends State<HomeTab> with AutomaticKeepAliveClientMixin {
+  int _plannedTransactionDays =
+      LocalPreferences.homeTabPlannedTransactionsDaysDefault;
+
   final TransactionFilter defaultFilter = TransactionFilter(
-    range: Moment.now()
-        .subtract(const Duration(days: 29))
-        .startOfDay()
-        .rangeTo(Moment.maxValue),
+    range: last30Days(),
   );
 
   late TransactionFilter currentFilter = defaultFilter.copyWithOptional();
 
-  final DateTime startDate =
-      Moment.now().subtract(const Duration(days: 29)).startOfDay();
+  TransactionFilter get currentFilterWithPlanned {
+    final Moment plannedTransactionTo =
+        Moment.now().add(Duration(days: _plannedTransactionDays)).endOfDay();
+
+    if (currentFilter.range != null &&
+        currentFilter.range!.contains(Moment.now()) &&
+        !currentFilter.range!.contains(plannedTransactionTo)) {
+      return currentFilter.copyWithOptional(
+          range: Optional(CustomTimeRange(
+              currentFilter.range!.from, plannedTransactionTo.endOfDay())));
+    }
+
+    return currentFilter;
+  }
 
   late final bool noTransactionsAtAll;
 
@@ -39,6 +54,17 @@ class _HomeTabState extends State<HomeTab> with AutomaticKeepAliveClientMixin {
   void initState() {
     super.initState();
     noTransactionsAtAll = ObjectBox().box<Transaction>().count(limit: 1) == 0;
+    LocalPreferences()
+        .homeTabPlannedTransactionsDays
+        .addListener(_updatePlannedTransactionDays);
+  }
+
+  @override
+  void dispose() {
+    LocalPreferences()
+        .homeTabPlannedTransactionsDays
+        .removeListener(_updatePlannedTransactionDays);
+    super.dispose();
   }
 
   @override
@@ -46,7 +72,10 @@ class _HomeTabState extends State<HomeTab> with AutomaticKeepAliveClientMixin {
     super.build(context);
 
     return StreamBuilder<List<Transaction>>(
-      stream: currentFilter.queryBuilder().watch(triggerImmediately: true).map(
+      stream: currentFilterWithPlanned
+          .queryBuilder()
+          .watch(triggerImmediately: true)
+          .map(
             (event) => event.find().search(currentFilter.searchData),
           ),
       builder: (context, snapshot) {
@@ -119,6 +148,13 @@ class _HomeTabState extends State<HomeTab> with AutomaticKeepAliveClientMixin {
         future: !range.from.isPast,
       ),
     );
+  }
+
+  void _updatePlannedTransactionDays() {
+    _plannedTransactionDays =
+        LocalPreferences().homeTabPlannedTransactionsDays.get() ??
+            LocalPreferences.homeTabPlannedTransactionsDaysDefault;
+    setState(() {});
   }
 
   @override