+++ lib/widgets/select_time_range_mode_sheet.dart
@@ -6,12 +6,17 @@ import 'package:go_router/go_router.dart';
 import 'package:material_symbols_icons/symbols.dart';
 
 enum TimeRangeMode {
-  thisWeek,
-  thisMonth,
-  thisYear,
-  byMonth,
-  byYear,
-  custom,
+  last30Days("last30Days"),
+  thisWeek("thisWeek"),
+  thisMonth("thisMonth"),
+  thisYear("thisYear"),
+  byMonth("byMonth"),
+  byYear("byYear"),
+  custom("custom");
+
+  final String value;
+
+  const TimeRangeMode(this.value);
 }
 
 class SelectTimeRangeModeSheet extends StatelessWidget {
@@ -52,6 +57,10 @@ class SelectTimeRangeModeSheet extends StatelessWidget {
               spacing: 12.0,
               runSpacing: 8.0,
               children: [
+                ActionChip(
+                  label: Text("tabs.stats.timeRange.last30days".t(context)),
+                  onPressed: () => context.pop(TimeRangeMode.last30Days),
+                ),
                 ActionChip(
                   label: Text("tabs.stats.timeRange.thisWeek".t(context)),
                   onPressed: () => context.pop(TimeRangeMode.thisWeek),