+++ lib/widgets/transaction_filter_head/transaction_filter_chip.dart
@@ -2,6 +2,7 @@ import 'package:flow/data/transactions_filter.dart';
 import 'package:flow/entity/account.dart';
 import 'package:flow/entity/category.dart';
 import 'package:flow/l10n/extensions.dart';
+import 'package:flow/widgets/utils/time_and_range.dart';
 import 'package:flutter/material.dart';
 import 'package:moment_dart/moment_dart.dart';
 
@@ -73,6 +74,10 @@ class TransactionFilterChip<T> extends StatelessWidget {
     }
 
     if (value case TimeRange timeRange) {
+      if (timeRange == last30Days()) {
+        return "tabs.stats.timeRange.last30days".t(context);
+      }
+
       return timeRange.format();
     }