+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Beta 0.6.0
+
+* Updated theme to correct `activeColor` for radio/checkboxes and its lists
+* Added filters to home tab, and added preferences for home page
+* Added error builder for Image `FlowIcon`s when the image is missing
+
 ## Beta 0.5.5
 
 * Selecting icons should be slightly better